+++ actix-service/CHANGES.md
@@ -2,7 +2,7 @@
 
 ## [1.0.0-alpha.3] - 2019-12-xx
 
-### Add missing Clone impl for factory_fn_cfg
+### Add missing Clone impls
 
 ## [1.0.0-alpha.2] - 2019-12-02